+++ Booster/Booster.xcodeproj/xcshareddata/xcschemes/Booster.xcscheme
@@ -129,6 +129,26 @@
                ReferencedContainer = "container:Booster.xcodeproj">
             </BuildableReference>
          </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "FAB252BA2753C5EB00A4FC5B"
+               BuildableName = "CoordinateTests.xctest"
+               BlueprintName = "CoordinateTests"
+               ReferencedContainer = "container:Booster.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "FAB252C82753D1AF00A4FC5B"
+               BuildableName = "MilestoneTests.xctest"
+               BlueprintName = "MilestoneTests"
+               ReferencedContainer = "container:Booster.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
